mserver file split
