+++ server/easyun/__init__.py
@@ -45,10 +45,10 @@ def create_app(run_env):
 def register_blueprints(app):
     """Register Flask blueprints."""
     from .common import auth
-    # from .modules import mserver
+    from .modules import mserver
 
     app.register_blueprint(auth.bp)
-    # app.register_blueprint(mserver.bp)
+    app.register_blueprint(mserver.bp)
     return None
 
 def configure_logger(app):
